[FIX] some comments and tried player death client side but did not succeed

+++ rtype_game/include/Game.hpp
@@ -101,7 +101,7 @@ class Game {
     RealEngine::ParallaxSystem        _parallaxSystem;
     rtype::Controls                   _controls;
 
-    std::unordered_map<int, std::shared_ptr<RealEngine::Entity>>      _players;
+    std::unordered_map<long int, std::shared_ptr<RealEngine::Entity>>  _players;
     std::unordered_map<long int, std::shared_ptr<RealEngine::Entity>> _entities;
     std::unordered_map<std::string, std::shared_ptr<sf::Texture>>     _textures;
     RealEngine::Sprite                                                _upSpaceship;
